fix: 매크로 잠금 시 재입력 멘트 미출력 수정 / totalDealtDmg int 오버플로우 수정
- LOCKED_YN=1 상태에서 재입력 시 빈 문자열 반환하던 문제 제거 → 매번 안내 멘트 출력
- OngoingBattle.totalDealtDmg int → long 변경 (ORA-17026 숫자 오버플로우 방지)
- monHpRemain 대입부 clamp 처리로 기존 int 사용처 변경 최소화

+++ src/main/java/my/prac/core/game/dto/OngoingBattle.java
@@ -10,7 +10,7 @@ import java.io.Serializable;
  */
 public  class OngoingBattle {
 	public int monNo;
-    public int totalDealtDmg;
+    public long totalDealtDmg;
     public Integer beforeJobSkillYn;
 
     // 진행 중 전투(now_yn=1) 묶음에서 ‘가장 최근’ lucky 상태
@@ -18,7 +18,7 @@ public  class OngoingBattle {
     
 
     public OngoingBattle setMonNo(int v){ this.monNo=v; return this; }
-    public OngoingBattle setTotalDealtDmg(int v){ this.totalDealtDmg=v; return this; }
+    public OngoingBattle setTotalDealtDmg(long v){ this.totalDealtDmg=v; return this; }
     public OngoingBattle setLuckyYn(Integer v){ this.luckyYn=v; return this; }
     public OngoingBattle setBeforeJobSkillYn(int v){ this.beforeJobSkillYn=v; return this; }
 }
